Put attribute_default test back in attribute/mod.rs
